Move tests into test/src
