+++ test/index-webgl-dependent-tests.js
@@ -1,8 +1,8 @@
 // Imports tests for all modules that depend on webgl
 
-import './webgl-utils';
-import './shadertools';
-import './webgl';
-import './core';
-import './experimental';
-import './deprecated';
+import './src/webgl-utils';
+import './src/shadertools';
+import './src/webgl';
+import './src/core';
+import './src/experimental';
+import './src/deprecated';
