+++ test/index-webgl-independent-tests.js
@@ -1,4 +1,4 @@
 // Imports tests for all modules that do not depend on WebGL
 
-import './utils';
-import './geometry';
+import './src/utils';
+import './src/geometry';
