+++ test/node.js
@@ -8,9 +8,10 @@ window.addEventListener = () => {};
 // Registers an alias for this module
 const path = require('path');
 const moduleAlias = require('module-alias');
-moduleAlias.addAlias('luma.gl', path.resolve('./src'));
+moduleAlias.addAlias('luma.gl/test', path.resolve(__dirname));
+moduleAlias.addAlias('luma.gl', path.resolve(__dirname, '../src'));
 
 // Run the tests
 require('./index-webgl-independent-tests');
 require('./index-webgl-dependent-tests');
-require('./debug/seer-integration');
+require('./src/debug/seer-integration');
