+++ test/src/core/index.js
