+++ test/src/io/read-texture.js
