+++ test/src/shadertools/lib/assemble-shaders.spec.js
