+++ test/src/webgl/buffer.spec.js
@@ -1,7 +1,7 @@
 import {GL, Buffer, isWebGL} from 'luma.gl';
 import test from 'tape-catch';
 
-import {fixture} from '../setup';
+import {fixture} from 'luma.gl/test/setup';
 
 test('WebGL#Buffer constructor/delete', t => {
   const {gl} = fixture;
