+++ test/src/webgl/helpers/query-manager.spec.js
@@ -1,5 +1,5 @@
 import test from 'tape-catch';
-import queryManager from '../../../src/webgl/helpers/query-manager';
+import queryManager from 'luma.gl/webgl/helpers/query-manager';
 
 test('WebGL helpers#queryManager', t => {
   t.ok(queryManager, 'Imported correctly');
