+++ test/src/webgl/query.spec.js
@@ -3,8 +3,8 @@
 import test from 'tape-catch';
 import {pollContext, Query} from 'luma.gl';
 import util from 'util';
-import {fixture} from '../setup';
-import GL from '../../src/webgl-utils/constants';
+import {fixture} from 'luma.gl/test/setup';
+import GL from 'luma.gl/webgl-utils/constants';
 
 function testQueryConstructDelete(gl, t) {
   const ext = gl.getExtension('EXT_disjoint_timer_query');
