+++ test/src/webgl/texture-cube.spec.js
